+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+* Support sorting IPAddrDTO
+
 ### Fixed
 * ProbeAlertDataDTO equivalence for times with different locations