+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [1.1.0] - 2016-05-27
 ### Added
 * Add terraform tags to structs to support mapstructure
 
@@ -11,7 +13,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * ProbeAlertDataDTO equivalence for times with different locations
 
 ### Changed
-* convert RawProfile to use mapstructure and structs instead of round-tripping through json
+* Convert RawProfile to use mapstructure and structs instead of round-tripping through json
 * CHANGELOG.md: fix link to v1.0.0 commit history
 
 ## [1.0.0] - 2016-05-11
@@ -20,4 +22,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * `Client` wraps common API access including OAuth, deferred tasks and retries
 
 [Unreleased]: https://github.com/Ensighten/udnssdk/compare/v1.0.0...HEAD
+[1.1.0]: https://github.com/Ensighten/udnssdk/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/Ensighten/udnssdk/compare/v0.0.0...v1.0.0